Conflicts:
	doc/help.txt
	e2e/create.spec.js
	package.json
	platforms.js
	spec/build.spec.js
	spec/compile.spec.js
	spec/config_parser.spec.js
	spec/emulate.spec.js
	spec/metadata/android_parser.spec.js
	spec/metadata/blackberry_parser.spec.js
	spec/metadata/firefoxos_parser.spec.js
	spec/metadata/ios_parser.spec.js
	spec/metadata/windows8_parser.spec.js
	spec/metadata/wp7_parser.spec.js
	spec/metadata/wp8_parser.spec.js
	spec/platform.spec.js
	spec/plugin.spec.js
	spec/prepare.spec.js
	spec/run.spec.js
	src/cli.js
	src/compile.js
	src/metadata/android_parser.js
	src/metadata/blackberry10_parser.js
	src/metadata/ios_parser.js
	src/metadata/windows8_parser.js
	src/metadata/wp7_parser.js
	src/metadata/wp8_parser.js
	src/platform.js
	src/plugin.js
	src/prepare.js
	src/run.js
	src/util.js

+++ src/prepare.js
@@ -17,6 +17,7 @@
     under the License.
 */
 var cordova_util      = require('./util'),
+    ConfigParser = require('./ConfigParser'),
     path              = require('path'),
     platforms         = require('../platforms'),
     platform          = require('./platform'),
@@ -33,7 +34,7 @@ var cordova_util      = require('./util'),
     util              = require('util');
 
 // Returns a promise.
-module.exports = function prepare(options) {
+exports = module.exports = function prepare(options) {
     var projectRoot = cordova_util.cdProjectRoot();
 
     if (!options) {
@@ -57,7 +58,7 @@ module.exports = function prepare(options) {
     var hooks = new hooker(projectRoot);
     return hooks.fire('before_prepare', options)
     .then(function() {
-        var cfg = new cordova_util.config_parser(xml),
+        var cfg = new ConfigParser(xml),
 	    internalDev = config.internalDev(projectRoot);
 
         // Iterate over each added platform
@@ -75,13 +76,19 @@ module.exports = function prepare(options) {
                 //or copy project config into platform if none exists
                 if (fs.existsSync(defaults_xml_path)) {
                     shell.cp("-f", defaults_xml_path, parser.config_xml());
-                    events.emit('log', 'Generating config.xml from defaults for platform "' + platform + '"');
+                    events.emit('verbose', 'Generating config.xml from defaults for platform "' + platform + '"');
                 } else {
                     if(fs.existsSync(parser.config_xml())){
                         shell.cp("-f", parser.config_xml(), defaults_xml_path);
                     }else{
                         shell.cp("-f",xml,parser.config_xml());
                     }
+                }
+
+                var stagingPath = path.join(platformPath, '.staging');
+                if (fs.existsSync(stagingPath)) {
+                    events.emit('log', 'Deleting now-obsolete intermediate directory: ' + stagingPath);
+                    shell.rm('-rf', stagingPath);
                 }
 
                 var platform_www = path.join(platformPath, 'platform_www');
@@ -95,17 +102,19 @@ module.exports = function prepare(options) {
                     shell.cp('-f', parser.cordovajs_path(libDir), path.join(platform_www, 'xface.js'));
                 }
 
-                // 提前用<project_dir>/www/config.xml中pre_install_packages配置覆盖平台工程config.xml中对应配置项
+                // 提前用<project_dir>/config.xml中pre_install_packages配置覆盖平台工程config.xml中对应配置项
                 // 后面的操作用依赖于该配置
-                var platformConfig = parser.config_xml(),
-                    platformConfigDoc = xml_helpers.parseElementtreeSync(platformConfig);
-                var newPackagesTag = et.XML(et.tostring(cfg.doc.find('./pre_install_packages'), {indent:4, xml_declaration:false}));
-                var root = platformConfigDoc.getroot();
-                if(platformConfigDoc.find('./pre_install_packages')) {
-                    root.remove(0, platformConfigDoc.find('./pre_install_packages'));
+                if(fs.existsSync(xml) && cfg.doc.find('./pre_install_packages')) {
+                    var platformConfig = parser.config_xml(),
+                        platformConfigDoc = xml_helpers.parseElementtreeSync(platformConfig);
+                    var newPackagesTag = et.XML(et.tostring(cfg.doc.find('./pre_install_packages'), {indent:4, xml_declaration:false}));
+                    var root = platformConfigDoc.getroot();
+                    if(platformConfigDoc.find('./pre_install_packages')) {
+                        root.remove(0, platformConfigDoc.find('./pre_install_packages'));
+                    }
+                    root.append(newPackagesTag);
+                    fs.writeFileSync(platformConfig, platformConfigDoc.write({indent:4}), 'utf-8');
                 }
-                root.append(newPackagesTag);
-                fs.writeFileSync(platformConfig, platformConfigDoc.write({indent:4}), 'utf-8');
 
                 // Replace the existing web assets with the app master versions
                 parser.update_www();
@@ -113,40 +122,17 @@ module.exports = function prepare(options) {
                 // Call xplugin --prepare for this platform. sets up js-modules appropriately.
                 var plugins_dir = path.join(projectRoot, 'plugins');
                 events.emit('verbose', 'Calling xplugin.prepare for platform "' + platform + '"');
-                xplugin.prepare(platformPath, platform, plugins_dir);
+                xplugin.prepare(platformPath, platform, plugins_dir, path.join(platformPath, '.xstaging'));
 
                 // Make sure that config changes for each existing plugin is in place
-                var plugins = cordova_util.findPlugins(plugins_dir),
-                    platform_json = xplugin.config_changes.get_platform_json(plugins_dir, platform);
-                if (plugins && Array.isArray(plugins)) {
-                    var plugman_cache = {};
-                    plugins.forEach(function(plugin_id) {
-                        if (platform_json.installed_plugins[plugin_id]) {
-                            events.emit('verbose', 'Ensuring plugin "' + plugin_id + '" is installed correctly...');
-                            xplugin.config_changes.add_plugin_changes(
-                                platform, platformPath, plugins_dir, plugin_id,
-                                /* variables for plugin */ platform_json.installed_plugins[plugin_id],
-                                /* top level plugin? */ true,
-                                /* should increment config munge? cordova-cli never should, only plugman */ false,
-                                plugman_cache
-                            );
-                        } else if (platform_json.dependent_plugins[plugin_id]) {
-                            events.emit('verbose', 'Ensuring plugin "' + plugin_id + '" is installed correctly...');
-                            xplugin.config_changes.add_plugin_changes(
-                                platform, platformPath, plugins_dir, plugin_id,
-                                /* variables for plugin */ platform_json.dependent_plugins[plugin_id],
-                                /* top level plugin? */ false,
-                                /* should increment config munge? cordova-cli never should, only plugman */ false,
-                                plugman_cache
-                            );
-                        }
-                        events.emit('verbose', 'Plugin "' + plugin_id + '" is good to go.');
-                    });
-                }
+                var munger = new xplugin.config_changes.PlatformMunger(platform, platformPath, plugins_dir);
+                munger.reapply_global_munge();
+                munger.save_all();
 
-                //Update platform config.xml based on top level config.xml
-                var platform_cfg = new cordova_util.config_parser(parser.config_xml());
-                platform_cfg.merge_with(cfg, platform, true);
+                // Update platform config.xml based on top level config.xml
+                var platform_cfg = new ConfigParser(parser.config_xml());
+                exports._mergeXml(cfg.doc.getroot(), platform_cfg.doc.getroot(), platform, true);
+                platform_cfg.write();
 
                 return parser.update_project(cfg);
             }).fail(function(e) {
@@ -157,3 +143,71 @@ module.exports = function prepare(options) {
         });
     });
 };
+
+var BLACKLIST = ["platform"];
+var SINGLETONS = ["content", "author"];
+function mergeXml(src, dest, platform, clobber) {
+    if (BLACKLIST.indexOf(src.tag) === -1) {
+        //Handle attributes
+        Object.getOwnPropertyNames(src.attrib).forEach(function (attribute) {
+            if (clobber || !dest.attrib[attribute]) {
+                dest.attrib[attribute] = src.attrib[attribute];
+            }
+        });
+        //Handle text
+        if (src.text && (clobber || !dest.text)) {
+            dest.text = src.text;
+        }
+        //Handle platform
+        if (platform) {
+            src.findall('platform[@name="' + platform + '"]').forEach(function (platformElement) {
+                platformElement.getchildren().forEach(mergeChild);
+            });
+        }
+
+        //Handle children
+        src.getchildren().forEach(mergeChild);
+
+        function mergeChild (srcChild) {
+            var srcTag = srcChild.tag,
+                destChild = new et.Element(srcTag),
+                foundChild,
+                query = srcTag + "",
+                shouldMerge = true;
+
+            if (BLACKLIST.indexOf(srcTag) === -1) {
+                if (SINGLETONS.indexOf(srcTag) !== -1) {
+                    foundChild = dest.find(query);
+                    if (foundChild) {
+                        destChild = foundChild;
+                        dest.remove(0, destChild);
+                    }
+                } else {
+                    //Check for an exact match and if you find one don't add
+                    Object.getOwnPropertyNames(srcChild.attrib).forEach(function (attribute) {
+                        query += "[@" + attribute + '="' + srcChild.attrib[attribute] + '"]';
+                    });
+                    foundChild = dest.find(query);
+                    if (foundChild && textMatch(srcChild, foundChild)) {
+                        destChild = foundChild;
+                        dest.remove(0, destChild);
+                        shouldMerge = false;
+                    }
+                }
+
+                mergeXml(srcChild, destChild, platform, clobber && shouldMerge);
+                dest.append(destChild);
+            }
+        }
+
+        function textMatch(elm1, elm2) {
+            var text1 = elm1.text ? elm1.text.replace(/\s+/, "") : "",
+                text2 = elm2.text ? elm2.text.replace(/\s+/, "") : "";
+            return (text1 === "" || text1 === text2);
+        }
+    }
+}
+
+// Expose for testing.
+exports._mergeXml = mergeXml;
+
